Fix Pydantic v2 validation error: Change LEADERBOARD_UPDATE_INTERVAL_SECONDS from int to float
Issue: Pydantic v2 is stricter than v1 and rejects int fields with fractional float values.
The LEADERBOARD_UPDATE_INTERVAL_SECONDS was typed as 'int' but had value '2.5'.

Error: Input should be a valid integer, got a number with a fractional part
[type=int_from_float, input_value=2.5, input_type=float]

Solution: Changed field type from int to float on line 93 since the research-backed
optimal value is 2.5 seconds for best UX (not too spammy, feels real-time).

Pydantic v2 Breaking Change:
- V1: Allowed float → int conversion (with silent data loss)
- V2: Only allows float → int if decimal part is 0 (e.g., 10.0 ✅, 10.5 ❌)

Tested:
- File compiles successfully
- quiz_settings loads correctly
- All backend files compile without errors

Python Version: 3.10
Pydantic Version: v2
Date: November 2025

+++ ata-backend/app/core/quiz_config.py
@@ -90,7 +90,7 @@ class QuizSettings(BaseSettings):
 
     # Batch leaderboard updates interval (seconds)
     # Research: 2-3 seconds is optimal for UX (not too spammy, feels real-time)
-    LEADERBOARD_UPDATE_INTERVAL_SECONDS: int = 2.5
+    LEADERBOARD_UPDATE_INTERVAL_SECONDS: float = 2.5
 
     # Number of top participants to show on leaderboard
     LEADERBOARD_TOP_COUNT: int = 5
